# backend/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Valkyris API
  version: 2.0.0
  description: Local-first home camera monitoring API. JSON responses use a success/message/data envelope; every response also includes X-Valkyris-Success and X-Valkyris-Message headers.
servers:
  - url: https://valkyris.local:8443/api/v1
security:
  - bearerAuth: []
paths:
...
  /admin/bootstrap:
    post:
      security: []
      summary: Create the first administrator account and device
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '201': {description: Administrator created, content: {application/json: {schema: {$ref: '#/components/schemas/APIResponse'}}}}
        '409': {description: Administrator already configured}
...
  /login:
    post:
      security: []
      summary: Authenticate a personal account by username and password
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '201': {description: Authenticated, content: {application/json: {schema: {$ref: '#/components/schemas/APIResponse'}}}}
        '401': {description: Invalid credentials}
  /pair:
    post:
      security: []
      summary: Register a new account using a single-use invitation; existing accounts must use login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PairRequest'}
      responses:
        '201': {description: Paired, content: {application/json: {schema: {$ref: '#/components/schemas/APIResponse'}}}}
...
  /me/credentials:
    post:
      summary: Set credentials for the authenticated legacy account or change existing credentials
      description: Initial setup requires the existing device session. Subsequent changes require currentPassword. Preserves the user ID and history, and revokes other sessions.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AccountCredentials'}}}}
      responses:
        '200': {description: Credentials saved}
        '400': {description: Invalid password or username unavailable}
        '401': {description: Authentication required}
  /me/password:
    post:
      summary: Change the current account password and revoke its other sessions
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChangePassword'}}}}
      responses: {'200': {description: Password changed, content: {application/json: {schema: {$ref: '#/components/schemas/APIResponse'}}}}}
...
components:
  securitySchemes:
    viewerCookie:
      type: apiKey
      in: cookie
      name: __Host-valkyris-viewer
      description: Secure HttpOnly SameSite=Strict cookie; requires X-Valkyris-Viewer 1 for all cookie-authenticated API calls. Read-only except WHEP negotiation and logout.
    bearerAuth: {type: http, scheme: bearer}
...
  schemas:
    APIResponse:
      type: object
      required: [success, message, data]
      properties:
        success: {type: boolean, const: true}
        message: {type: string}
        data: {}
...
    ChangePassword:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: {type: string, format: password}
        newPassword: {type: string, format: password}
...
    AccountCredentials:
      type: object
      required: [newPassword]
      properties:
        username: {type: string, minLength: 3, maxLength: 40}
        currentPassword: {type: string, format: password, writeOnly: true}
        newPassword: {type: string, format: password, minLength: 12, writeOnly: true, description: At most 72 UTF-8 bytes}
...
    LoginRequest:
      type: object
      required: [username, password, deviceName]
      properties:
        username: {type: string, minLength: 3, maxLength: 40}
        userName: {type: string, description: Display name used only during account creation}
        browserAdmin: {type: boolean, description: Request an HttpOnly browser session. Permissions always come from the authenticated account.}
        password: {type: string, format: password}
        deviceName: {type: string}
        readOnly: {type: boolean, default: false, description: "Issue a renewable 30-day HttpOnly cookie session; requires X-Valkyris-Viewer: 1; no token is returned in JSON."}
        locale: {type: string, default: pt-BR}
    PairRequest:
      type: object
      required: [code, deviceName, username, password]
      properties:
        username: {type: string, minLength: 3, maxLength: 40}
        password: {type: string, format: password, minLength: 12, writeOnly: true, description: At most 72 UTF-8 bytes}
        userName: {type: string, description: Display name}
        code: {type: string}
        deviceName: {type: string}
        locale: {type: string, default: pt-BR}
...